+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@
 
 * #562 Column getColumnName() and toQL() consistency
 * #563 QL: "startsWith(..)" function without the "str(..)" wrapper for the first argument
+* #564 QL: lower(..), upper(..) without the "str(..)" wrapper for the argument
 
 ## Release 2.0.0-M5